fix!: change EnvType variants: 0-int, 1-float, 2-string (#61)
According to DBC documentation:

```
env_var_type = '0' | '1' | '2' ; (* 0=integer, 1=float, 2=string *)
```

P.S. Also disable codecov patch minimum - it should be informative, not
CI crashing

+++ src/parser_tests.rs
@@ -213,7 +213,7 @@ EV_ IUV: 0 [-22|20] "mm" 3 7 DUMMY_NODE_VECTOR0 VECTOR_XXX;
 "#;
     let exp = EnvironmentVariable {
         name: "IUV".to_string(),
-        typ: EnvType::Float,
+        typ: EnvType::Integer,
         min: -22,
         max: 20,
         unit: "mm".to_string(),
